+++ changelog.md
@@ -2,10 +2,10 @@
 
 All notable changes to `TestAssist` will be documented in this file.
 
-## Version 1.1
+## Version 1.x
 
 ### Added
-- Assertions for whether a model is inserted or updated in the database.
+- Assertions for whether a model is inserted, updated or deleted in the database.
 - Moved `FabricateModels` to the same `Database` namespace.
 
 ## Version 1.0